requirements.txt
web3 
kafka-python 
python-dotenv
pytest
requests
numpy
pandas
parquet
pyarrow
duckdb
dbt-core 
dbt-duckdb
streamlit
altair
streamlit_autorefresh